Added text2cypher agent, reorg folders, and fixed setup
